+++ demos/postcss.config.js
@@ -1,10 +1,6 @@
-import tailwind from 'tailwindcss'
-import autoprefixer from 'autoprefixer'
-import tailwindConfig from './tailwind.config.js'
-
-export default {
-  plugins: [
-    tailwind(tailwindConfig),
-    autoprefixer,
-  ],
+module.exports = {
+  plugins: {
+    tailwindcss: {},
+    autoprefixer: {},
+  },
 }