+++ packages/react/src/useReactNodeView.ts
@@ -2,7 +2,7 @@ import { createContext, useContext } from 'react'
 
 export interface ReactNodeViewContextProps {
   onDragStart: (event: DragEvent) => void,
-  maybeMoveContentDOM: () => void,
+  nodeViewContentRef: (element: HTMLElement | null) => void,
 }
 
 export const ReactNodeViewContext = createContext<Partial<ReactNodeViewContextProps>>({