+++ packages/extension-highlight/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [2.0.0-beta.20](https://github.com/ueberdosis/tiptap/compare/@tiptap/extension-highlight@2.0.0-beta.19...@tiptap/extension-highlight@2.0.0-beta.20) (2021-09-29)
+
+**Note:** Version bump only for package @tiptap/extension-highlight
+
+
+
+
+
 # [2.0.0-beta.19](https://github.com/ueberdosis/tiptap/compare/@tiptap/extension-highlight@2.0.0-beta.18...@tiptap/extension-highlight@2.0.0-beta.19) (2021-09-28)
 
 **Note:** Version bump only for package @tiptap/extension-highlight